Improved the notification screen with users and fast serch. Made access right check lenient for ComponentPropertyEditor.EditProperty so you don't have to access right control all properties. Right click is now inactive when ckeditor is active.

+++ src/java/org/infoglue/cms/controllers/kernel/impl/simple/PublicationController.java
@@ -852,34 +852,37 @@ public class PublicationController extends BaseController
 
 		try
 		{
-		    for(EventVO event : events)
-		    {
-	    		List<Event> repoEvents = EventController.getPublicationEventVOListForRepository(event.getRepositoryId(), db);
-			    for(Event repoEvent : repoEvents)
+			if(!isDeleteOperation)
+			{
+			    for(EventVO event : events)
 			    {
-			    	if(!event.getId().equals(repoEvent.getId()) && event.getEntityClass().equals(repoEvent.getEntityClass()))
-		   			{
-				    	if(event.getEntityClass().equals(ContentVersion.class.getName()))
-				    	{
-				    		ContentVersionVO eventCVVO = ContentVersionController.getContentVersionController().getContentVersionVOWithId(event.getEntityId(), db);
-				    		ContentVersionVO repoEventCVVO = ContentVersionController.getContentVersionController().getContentVersionVOWithId(repoEvent.getEntityId(), db);
-				    		if(eventCVVO.getContentId().equals(repoEventCVVO.getContentId()) && eventCVVO.getLanguageId().equals(repoEventCVVO.getLanguageId()))
-				    		{
-				    			EventController.delete(repoEvent, db);
-				    		}
-				    	}
-				    	else if(event.getEntityClass().equals(SiteNodeVersion.class.getName()))
-				    	{
-				    		SiteNodeVersionVO eventSNVVO = SiteNodeVersionController.getController().getSiteNodeVersionVOWithId(event.getEntityId(), db);
-				    		SiteNodeVersionVO repoEventSNVVO = SiteNodeVersionController.getController().getSiteNodeVersionVOWithId(repoEvent.getEntityId(), db);
-				    		if(eventSNVVO.getSiteNodeId().equals(repoEventSNVVO.getSiteNodeId()))
-				    		{
-				    			EventController.delete(repoEvent, db);
-				    		}
-				    	}
+		    		List<Event> repoEvents = EventController.getPublicationEventVOListForRepository(event.getRepositoryId(), db);
+				    for(Event repoEvent : repoEvents)
+				    {
+				    	if(!event.getId().equals(repoEvent.getId()) && event.getEntityClass().equals(repoEvent.getEntityClass()))
+			   			{
+					    	if(event.getEntityClass().equals(ContentVersion.class.getName()))
+					    	{
+					    		ContentVersionVO eventCVVO = ContentVersionController.getContentVersionController().getContentVersionVOWithId(event.getEntityId(), db);
+					    		ContentVersionVO repoEventCVVO = ContentVersionController.getContentVersionController().getContentVersionVOWithId(repoEvent.getEntityId(), db);
+					    		if(eventCVVO.getContentId().equals(repoEventCVVO.getContentId()) && eventCVVO.getLanguageId().equals(repoEventCVVO.getLanguageId()))
+					    		{
+					    			EventController.delete(repoEvent, db);
+					    		}
+					    	}
+					    	else if(event.getEntityClass().equals(SiteNodeVersion.class.getName()))
+					    	{
+					    		SiteNodeVersionVO eventSNVVO = SiteNodeVersionController.getController().getSiteNodeVersionVOWithId(event.getEntityId(), db);
+					    		SiteNodeVersionVO repoEventSNVVO = SiteNodeVersionController.getController().getSiteNodeVersionVOWithId(repoEvent.getEntityId(), db);
+					    		if(eventSNVVO.getSiteNodeId().equals(repoEventSNVVO.getSiteNodeId()))
+					    		{
+					    			EventController.delete(repoEvent, db);
+					    		}
+					    	}
+					    }
 				    }
 			    }
-		    }
+			}
 		}
 		catch (Exception e) 
 		{
